Moved documentService to own Service folder
